CPack: Add a debian-revision and maintainer info
From man deb-version: It is conventional to restart the debian-revision
at '1' each time time the upstream-version is increased.

Not really required, since "no revision" sorts earlier than "some
revision", but seems to be conventional.

Also add a maintainer string, that is shown in the package info and
marks this package to come from the "Qt Project" directly.

Change-Id: I417a8d14d82f9f8ccfbe791e767d4ffed8175604
Reviewed-by: Cristian Adam <cristian.adam@qt.io>