Android: Launch "adb track-devices" with explicit working directory
Since no working directory was explicitly set, the "bin/" directory of
QtC was chosen as working directory.

"adb track-devices" leaves the server process alive after terminating,
and that server process used QtC's working directory. On Windows, that
caused that directory to be undeletable/unmovable as long as the adb
server was alive.

This change sets the working directory of "adb track-devices" to the
parent dir of the adb binary.

Task-number: QTCREATORBUG-28932
Change-Id: I3adf52829c9f4bf456d05d2685c04314e1d25a40
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>