McuSupport: Add dialog for Qt MCUs kit creation status and help
The status of automatically created kits or manually updated ones is
displayed in "General Messages" along with other messages making it
difficult to know if something went wrong with kit creation.

To avoid showing a dialog at QtC startup an InfoBar message was added
with a detail button that will open a dialog listing all the messages
with their platform, package and Os providing an easy button to go to
the MCU Option page to fix the package of a help button pointing to the
prerequisites documentation.

Fixes: QTCREATORBUG-28920
Change-Id: I20c30d968fbd2d4c6d10d16504a51e7201a2db3b
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>