HelperWidgets: Explicitly import Basic style
HorizontalScrollBar and VerticalScrollBar customizes ScrollBar. The
default Controls style used in Qt 6 depends on the platform the
application runs on, and in case of Windows, it will use the native
Windows style. The Windows style is not customizable. To avoid issues
(like warnings about transitionDuration not being defined), explicitly
opt into the Basic style, which can be customized.

Fixes: QTBUG-107771
Change-Id: I141b3466964e5c3c64b39ca73a85eac14b9b6202
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Henning Gründl <henning.gruendl@qt.io>