Scxmleditor: Add showing onEntry and on Exit events
This commit enhances the SCXML Editor by adding a new
feature that enables the display of onEntry and onExit
events in the state item. This feature provides developers
with a more comprehensive view of the state machine's behavior,
allowing them to better understand the transitions that occur
when entering or exiting a particular state.

Fixes: QTCREATORBUG-17378
Change-Id: Iaf84af209d9acf13bd4183f7da9acbcd6c1c8214
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>