ResultsCollector: Fix destructor
It may happen that still running LocatorMatcher is destructed
from the LocatorWidget destructor (i.e. from CorePlugin d'tor).
In this case the global future synchronizer is already reset,
so the d'tor of ResultsCollector can't put running deduplicator
task into the global synchronizer.

Check this case and place the deduplicator task into the global
future synchronizer only when it still exists. Otherwise, blocking
wait for the deduplicator to finish.

Change-Id: Ie788db923808dd91bdfdc4f1f53d336baf7c672a
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>