Utils: Introduce QTC_SCOPED_TIMER()
QTC_SCOPED_TIMER() is a handly debug tool that measures
the time spent in a block of code. It starts measuring the time
when the line where it was placed is reached, and stops measuring
the time when the current invocation exits the block in which it
was placed (i.e. when QTC_SCOPED_TIMER() goes out of scope).

The QTC_SCOPED_TIMER does two printouts:
1. When it starts - it prints the current time, filename and
   line number.
2. When it ends - it prints the current time, filename,
   line number and the timeout in ms.

The QTC_SCOPED_TIMER() was added into qtcassert.h file on
purpose, as this file is included already in most of the
codebase. In this way, when it needs to be used, it's enough
to add a QTC_SCOPED_TIMER() without adding extra #include.

Example use case, after adding the "QTC_SCOPED_TIMER()" into
ProjectExplorerPlugin::initialize() as a first line:
SCOPED TIMER [14:46:57.959] in [_long_path_here_]/
projectexplorer.cpp:823 started
SCOPED TIMER [14:46:58.087] in [_long_path_here_]/
projectexplorer.cpp:823 stopped with timeout: 127ms

Change-Id: Iaed3f297c8aeb6e90dd9909e76fc9933599a39b6
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>