MultiTextCursor: iterate over map entries
This will ensure that the cursors are iterated over in the order they
appear in the document. This fixes an issue on copy and paste a
multitext cursors, since the text was copied in the order inside the
document, but inserted in the order how the different cursors were added
to the multi text cursor.

Fixes: QTCREATORBUG-29117
Change-Id: I8475cde7129c378ec7862482226db25e40f61e1b
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>