Show qmllint errors on issues pane
Some warning / error messages of qmlsc/qmllint tools were not handled
by any of the output parsers and hence no CompileTask were created to
handle those output, like pushing those warnings into Issues pane. This
was also causing incorrect formatting of CompileOutput messages.This is
not an issue for MsvcParser because there a fallback task is created for
"Error" and "Warning" lines.

Introduce a fallback mechanism in QtParsers to handle qmltooling output
and add test.

Fixes: QTCREATORBUG-28720
Change-Id: I6d0de6bc8b6869c6ea80b72018a4acb7decc8b6d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>