Fix issues with project tree node expansion
In CMake projects, initially the project name is the directory name
that is later changed to the name defined in the project file. If
the user had expanded the project when it was still in its original
name, it would be left in the expand data as is, causing the project
to be expanded every time it is re-opened.

Fixed by checking if the name had changed and then re-inserting
possible expanded data with the new name.

Fixes: QTCREATORBUG-28681
Change-Id: I3b914fd0ae916205c86a6bd79cab5d57499d4d33
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>