TaskTree tests: Get rid of future synchronizer
The future synchronizer may abuse the subsequent test invocation
as it may still have spinning threads from the previous test run.

As all the started tasks should finish quickly on cancel,
we drop putting the canceled futures into the synchronizer.
Instead, we blocking wait for them to finish.

In this way, when the subsequent test is invoked, we should expect
the all available threads are ready to be used.

Change-Id: I9bb022b8c0950d27cb323ff83bf18148829160b3
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>