Issues pane: Fix filtering
If the parent is visible, we always want to show the child as well.
The previous implementation would always assume the row number to be a
top-level entry, leading to seemingly random behavior if any sort of
filtering was active.
Amends 778d7a981995898a3afd782bfc199d032d4a80f0.

Change-Id: I354c1b07929c45034a7d24b4c31f986ddd01c877
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>