LanguageClient: postpone creating progressbars
Some language servers are spawning progress reports for many small
tasks. This can get distracting in the case where those progress bars
are spawned while typing, like in the case of the python language server
that creates a report for every lint that gets triggered after receiving
a document change notification. So in order to reduce the amount of
progress bars created from progress reports we postpone the creation of
the bars.

Fixes: QTCREATORBUG-29224
Change-Id: I2e658be0a26b21e41c80b444184648ba70682522
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>